requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
openai==0.28.1
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7